added macro LocalizedStringInThisBundle to iMedia.h updated genstrings build phase to look for LIB rather than NSLocalizedString
Discussion: Since this is a framework, calls to NSLocalizedString may not
work as that macro generally looks in the main (application) bundle for
the strings file. LocalizedStringInBundle fixes this by looking in the
bundle that contains the object code where the macro was called.

genstrings then uses -s LocalizedStringInThisBundle to convert the calls
properly and build a good Localizable.string file