Finally the solution to the scrollview nightmare.
Removed the prepare method from MUPhotoView and subsequently from the controllers

Added the Epeg framework - This is a Universal Binary framework.
To get the speed advantage of generating thumbs with this library you need to copy
the Epeg framework into your application wrapper. It gracefully falls back to sips
if you don't want to or forget to include it in the wrapper.