Restored the missing play button icons from french, zh_TW -- no idea why they were missing!
NOTE: Make sure these fixes don't get clobbered if/when I merge in new localizations.

Fixed a problem on list of links where the icon was causing the baseline of the text to be funny

Changed the time value trnasformer and music folder parser to assume that the time value is in milliseconds, not seconds.  I think that older versions of itunes must have stored their times in seconds, but now it's milliseconds.  I don't know if there is a way to distinguish between an old and new itunes library file, so I'll just assume people have the new itunes.

**** One issue outstanding that was reported to me -- a problem with the photo grid view now showing the last column properly.  I will send this with an example movie to the email list.