Big fixes to movie previews in the movie module.
Double-clicking on a thumbnail now just starts up the movie, finally.

Resizing of the thumbs or the windows now properly brings the movie back when the resizing is done.

(This required some changes to our modification of MUPhotoView -- at some point it would be nice to try to get our changes sent "upstream")


I took out a bunch of NSLogs that were too chatty