MUPhotoView - Major changes to click & drag handling.
There was a bug whereby if there was already a selection in the view and an unselected photo is clicked on and the user starts to drag, then this would be treated as a drag select sequence, instead of the expected drag & drop sequence. Fixing this bug lead to a simplification of the whole mouseDown/mouseDragged/mouseUp sequence and it now behaves like the iLife browsers. See the comment at the top of mouseDown for a better understanding.