Checking in updated MUPhotoView, based on MUPhotoView 1.2.  This probably needs a bit of pounding upon, but it seems pretty good.
Note that I've taken out the fading code for now, it wasn't really working.  Maybe we can get Blake to add it in later.


Renaming all "lproj" directories to their ISO language equivalents. This is because iLocalize (which we use for the localizations) says that Apple recommmends against mixing the old-school names (like English.lproj) with the ISO codes (like zh_TW), and it oh-so-helpfully converts them. So I'm making the change in the source repository.