Lots of little tweaks.
Updated project file to include headers where appropriate (framework)
and not include headers where inapropriate (test app)

Made MUPhotoView.h available to clients in iMedia.h.  We might want to
add other headers to this too, not sure.

Moved toolbar icon image setScalesWhenResized:YES to one place so it's
not duplicated all over the place

A bit more error checking on Ener/Exit movies on thread, just log
in case of an error so we can see if there was a problem

Allow the delegate and the parser to notice when the top outline view
has expanded, method iMediaBrowser:willExpandNode: is called.

Clean up the protocols a bit