Fixed a crasher and another potential crasher by creating/accessing QTMovie objects in a background thread from the iTunes parser and the GarageBand parser.  Calling EnterMoviesOnThread/ExitMoviesOnThread seems to do the trick.  (Can somebody with GarageBand verify that this is good?)
Also, I stumbled upon the fact that when previewing a music item that
actually was a video, it would start playing the video on the iMedia
window!  What I did was to disable any non-audio tracks, and not bother
playing if there are no audio tracks at all.

MINOR BUG -- ANYBODY KNOW OF A WORKAROUND?  If it's an "MPEG" file (which
I am assuming is a multiplexed audio/video file), then you don't get any
audio preview.  So to get around that, we'd need to come up with a better
way of disabling the video preview while letting the audio play.  Any ideas?

Fixed the table for the English music nib so it's better resized.  This
will eventually propagate to the other languages.