Progress on SplitView NSOutlineView/NSPopupButton swapping.  Some problems still.
Cleaned up several folder-parsing methods to deal with hidden directories. 
First I check if a file is visible, THEN make a decision on whether to process folders vs. files.

VERY preliminary work on caching movie previews as files, so the hit will be minimized.
(I've made a method to create a cache file path based on an SHA1 hash of the given file path.)

Added improved 'description' methods for DEBUG-only builds of the test application,
to help in debugging.



Still in progress:
I'm not seeing my "Movies" folder show up ... trying to diagnose why.  The parser is
finding files, and setting the array in the properties in the key of "Movies" but
then the logic does not care about that.  I'm trying to understand what's going on.