Restored the main class to only use the nil check as to whether or not to include a parser's top node.  This fixes a regression where items without subfolders weren't getting added.
Updated a lot of the parsers to return nil if there is no database.
Tested with just about all parsers.

Fixed a couple of leaks.  Moved some autoreleases up to nest along
with the initial allocation of objects, so that their lifetime management
didn't need to be further worried about.  Just a defensive style.

Removed the *old* Aperture parser; it was not used.

Added a very simple parser for ~/Library/Sounds.  It ought to merge in
the contents of the NSSearchPaths... but that would take longer.

Made the photo placeholder rounded-rect even subtler.