EPeg is gone, replaced by Core Graphics.  MUCH faster.  I could further speed things up by converting MUPhotoview to use the CGImageRef's but I didn't want to do anything drastic.
Cleaned up the cache lock a bit to make it clearer when it's used

Max thumbnail size for movies and images to be 240 pixels, which is the
slider's max.

New simple parsers to pick up other cool sources of images:
User Pictures, Desktop Pictures, Screensaver Pictures, iChat icons

Also added an old early PROTOTYPE spotlight parser -- NOT included in
the project, I thought I'd just check it in, in case somebody wants to work on it later.  (The idea is to use Spotlight to browse images, music, etc.
that aren't in the usual places, also maybe search.)

Put the typical header into the source files that didn't have 'em