Various leak fixes, mostly by deallocating ivars
Added a few minor utility methods to try and deallocate caches and such,
for leak testing.  A menu item in the tester app menu invokes one.

Minor tweaks in English nibs, like removing unused title of new button,
so that in the future it won't need to be localized.

Put in DEBUG wrappers to not initially select a browser.  The actual
built product auto-selects the first, or most recently used, browser.