Per pierre's discovery, took out release of mySelectedBrowser
We're now using NSURLCache to store movie thumbnails on disk.

Factored out the thumbnail (or placholder file icon) generation to work
from both foreground and background thread, see getPosterImageOrPlaceholderFromMovie:
We test if there is a dimension to the movie, returning NSNull as a placeholder if there is no actual image so we'll continue to show the file icon.


Note: I'd prefer a new adhoc URL scheme, but I was having problems, so
we're using a file:// URL of a pseudo file (the original file path, with
".tiff" appended).  Works fine.

Note: Also I wanted to
have our own iMedia-specific URL cache so apps could share the cache, but I
couldn't get the cache to ever be written to disk.  Using the shared URL cache,
it is written after ~15 seconds of idle time.