First patch from Alf Watt.
Applied via patch -p0 < imedia-changes-20070228.svndiff

Note that I took out the warning about the missing delegate method -- 
this is purposely not warning in order to discourage nasty companies from
just upgrading to this version of the imedia code without getting the
scoop on how to do it.  We will probably take out this restriction
once we get the flipping-window changes that have been forthcoming for
a while.