Added a bunch of useful metadata for tooltips from the photo and movie views.  The photo information comes from the iPhoto library and the photos themselves, including EXIF photo timestamp.  The movie information comes from the appropriate sources - iPhoto or iTunes, plus the movies themselves; this information is cached with the thumbnail when we have to generate the thumbnail ourselves.
Factored the iPhoto parser classes into a superclass since there is so much
in common.

MUPhotoView now calls a new delegate method to make it clear that it is
a tooltip we are asking for. Added a new 'showCaptions' property that client
can use to determine if captions should be used.  (At some point I want
to make this be something users can change)


Anyhow, please try out the tooltips and see if you like the result!


also --

Took out the debug-build optimization I had "temporarily" turned on, oops!