Patch from Airy ANDRE:
If you scroll upwards through a large library, if an image border  
happens to align with the scrollview border, the shadow draws  
incorrectly and stays like that until the mouse is released, because  
the image is completely clipped and then does not cast any shadow.

Scrolling more won't invalidate the shadow part since it's already on  
screen, so the shadow won't be correctly updated, until next full  
view update.

On scrolling, it should draw slightly more (according to shadow  
size), to make sure the shadow appear on the next step.