A few minor tweaks:
* New "protocol" for string passed to setIconName:  --

If you give it an app identifier with a colon (:) and an image name, it will
find the given image in the given app identifier - this to make it easy to not
have to copy image resources.

If you just pass in the app identifier ending in colon with no resource, you
will get the app icon.  If you still pass in just the app identifier that will
work too; don't want to break any existing code, but it will check that there
are at least two dots, e.g. com.mycompany.foo

This logic tries to be forgiving of missing image files, so that in case you have
an old library file around but not the application, or names have changed, you'll
get a generic application icon or folder icon.

TODO: remove all of the old iApp little images from iMedia so we can just
refer to those used in the iApps.

* Adjust the baseline of the icon in the source outline so it's not up against the top edge.  

TODO: How do we make the text baseline be up just a bit?

* Project: took out the build settings from the top level to avoid confusion; those
should all be overridden in the xcconfig files.

* Our subscription to the UK file watcher now ignores "attribute changed" events.
I was getting some of these spuriously; no need to trigger a reparse then.

* Minor change to the iMediaBrowserProtocol, "library:" now returns an id, which is normally
a iMBLibraryNode* but can also be an NSArray of iMBLibraryNode*'s.  EXISTING CODE should work
fine but you will want to update your declarations to remove andy compiler warnings.
(This will allow a parser to create > 1 top level node, useful for instance for > 1 aperture
library)