Applied patch from Martin Redington <martin@mildmanneredindustries.com>
Fixes 2 and 3 from the list below, and provides a hook of where to "fix" #1 -

Not commiting to that new behavior until we get some more feedback from people.




1) When there is only a single row of images, spacing is incorrect.

in the case with only one image, the image is centred in the view,
rather than being up against the left hand side.

with two images, they are evenly spaced across the view, instead of
occupying the two left most columns, and so on.

2) When all items are selected, cursor keys don't do anything.

Try this in the Finder. Select All. Hit any cursor key. You will be
lett with only one selected item (which one depends on which cursor
key you hit).

Try it in imedia's photoview. the cursor key has no effect.

3) Shadow is drawn when item is selected.

I don't think selected items should have shadows, but they do.