From Chris Meyer:
This patch fixes a problem where if the iPhoto library contains files that are aliases to images rather than the images itself. This can occur when the photos are not copied to the iPhoto library but instead are copied as aliases. This was the behavior for some older versions of iPhoto, but it may not be the behavior for the current version. In any case, my users have libraries where this resolves the problem (no pun intended!).