From Chris Meyer (minor adjustments to get MetaDataTool to link).  This looks good.
Here is a patch which improves the performance of reading the metadata. This patch is not for the feint of heart... threading isn't simple!

It would be best for someone to carefully go over this code. I did quite a bit of testing, but it's complicated code and could probably use another set of eyes.

I have introduced a new class 'MetadataUtility' which handles reading of metadata for music files. The class first tries to use the metadata provided by spotlight for the file. This is very fast. If the Spotlight metadata cannot be used, the class uses a secondary process 'MetadataTool' to read the QTMovie and extract the metadata as before. This is slower than spotlight, but still fast. Finally, if none of those work, it tries to use the main thread in the application itself to read the metadata. This is slow.

Since spotlight and the secondary process work almost all of the time, this takes all of the heavy lifting off the main thread and puts it into background threads in a safe manner. The response in the foreground thread should be great now.

This patch also fixes a bug I introduced earlier in that the subfolders of the various music directories were not properly updated. This was actually a KVC problem on 'items' and 'isLeaf' that I fixed along the way.

Specific changes:

'iMediaBrowser.m': Minor change for KVC coding of 'item' in iMBLibraryNode class.

'iMBLibraryNode.h': Minor change for KVC coding of 'item' in iMBLibraryNode class.

'iMBGarageBandParser.m': Minor change for KVC coding of 'item' in iMBLibraryNode class.

'iMBAbstractParser.m': Minor change for KVC coding of 'item' in iMBLibraryNode class.

'iMBiPhotoAbstractParser.m': Minor change for KVC coding of 'item' in iMBLibraryNode class.

'iMBAperturePhotosParser.m: Minor change for KVC coding of 'item' in iMBLibraryNode class.

'iMBLibraryNode.m': Made iMBLibraryNode KVC compatible for 'items' and 'isLeaf'.

'iMBMusicFolder.m': Changed metadata code to use new MetadataUtility class. Also fixed subfolder problem.

'iMediaBrowse.xcodeproj/project.pbxproj': Added new files MetadataUtility.h/m and MetadataTool.h/m. Also added new target for building auxiliary MetadataTool and made iMediaBrowser dependent on it.

'MetadataUtility.m', 'MetadataUtility.h': Class to handle reading of metadata from music files. Called from the application.

'MetadataTool.m', 'MetadataTool.h': Process to handle reading of metadata. Called from MetadataUtility.