Incorporated some updates from Pierre, for lightroom.
Pierre -- CAN YOU TEST PLEASE with lightroom?


I updated some of the API so that database parsing more explicitly returns an iMBLibraryNode* for parseDatabase, but still has a separate method to override if you want to return an array of nodes (for a parser to put multiple items at the top level -- like LightRoom)

Also fixed the Safari parser so that it handles bookmarks at the top level ... before it was assuming the top level was always a folder!