This is a replacement cache that actually works in Leopard.
MAJOR, HUGE KUDOS to Martin Wennerberg of norrkross.com for this patch.

This is ingenious -- it used distributed objects (which I've barely been able to get my head around) with a separate process to do the quicktime thumbnails.  It also has support [not yet implemented here] for 'mini-movies' to allow scrubbing through the movies.

This should increase stability quite a bit.  Also I see it's using JPEG 2000 for the thumbnail, which should be nice an optimial.  (Maybe we could make a user default for the file type used, in case the client app wants to use the thumbnail in a different format, like JPEG for web publishing)


(I've made a few minor tweaks to Martin's submission.)