From Daniel Jalkut:
A problem exists where if an item being displayed by MUPhotoView doesn't have supply a valid image, it can't be manipulated.  Since MUPhotoview displays a placeholder image in lieu of the nil image, I think it makes sense to treat that placeholder image as the image for hit testing of selection, dragging, etc.

This patch modifies the behavior of photoRectForIndex: and mouseDragged: to accomplish this. With the patch in place, "bogus" images can now be selected and manipulated as expected.