From Peter Baumgartner:
I have a bugfix which addresses issue 43. I modified iMBTunesMusicParser to accurately represent the hierarchical structure in iTunes. This also includes accurate 
modern icons. The attached file contains a diff for the sources, complete sources, some changed resources (strings files), and some new image resources (which 
need to be added to the project). I hope that nothing is missing. 

Here's a brief overview of the modifications:

1) Added some methods to NSWorkspace+iMedia and NSImage+iMedia which are needed by my modified parser classes.
2) Made sure that copyWithZone: in iMBLibraryNode also copies the myParser pointer
3) Rewrote iMBiTunesMusicParser to use modern icons and accurately represent hierarchical structure of the library.
4) Made sure that all iMBLibraryNodes that are created have their parser set. This is important in the new controller classes that I am currently writing.
5) Added 2 entries to the strings files. Attention, only added to en.lproj, de.lproj, and fr.lproj. Other localizations are still missing.
6) Added image resources for the modern icons. There are separate sets of icons for iTunes 7 and older versions of iTunes.
7) I did not touch any of the existing icons in the project, so that the old implementation can continue to be used.

At the top of iMBiTunesMusicParser.m there is a #define RECURSIVE_PARSEDATABASE which controls which code path is used. Define it to 1 to use my new 
implementation, or 0 for the old implementation. Hope you can get it to run and that you like it. Let me know if anything is missing.



NOTE: What do people think about keeping around the old iTunes 6 icons?  Is it worth the trouble?

Also, the Localizable.strings files didn't seem to be changed.  Let's see if we can try this again -- Peter maybe you can just send the strings in an email message?