From Peter B: Fix for issue 36
Hi guys,

Here's a fix for this issue. The important part was to check for value!=nil before calling [value rangeOfString:lowerSearch], or a nil value would lead to a positive 
match, and thus to way too many search results.

I also generalized searching a bit by adding a searchableProperties array and iterating across this array during search. This array is pre-initialized with 
{@"Name",@"Artist",@"Album",@"Genre"}, but it can easily be changed with the setter.

Cheers,
Peter