From Daniel Jalkut:
I noticed the badge that gets drawn when multiple items are dragged was clipped against the top edge of the drag image, when it was clear that it was intended to be inset some from the top edge. I  fixed this bug and while I was added made a number of improvements which I think make the badge drawing more readable and attractive.

Improvements include:

- Bolder, larger font for the number

- Darker, non-transparent red for the color shows better against white text

- Shadow applied behind the circle for subtle contrast with transparent image

- More accurate centering of the text in the circle

- The code is generally more robust now, so that if anybody wants to tweak or parameterize the sizes, it should accommodate you easily.

- The code includes code where Mail-like star shape is used instead of the plain circle. I can't decide yet whether this is better or not. I sort of feel like this star style of badge might be on the way out in favor of something cleaner that would be more in line with the clean styles of Leopard. However I currently think the star-style that I'm using is most appropriate for the current shipping examples in Apple's own code, so I left it macro-enabled by default in the patch.