Issue 57 from Peter B:
iPhoto nodes should be displayed with the current icons (i.e. iPhoto 7 at the moment). 

Nodes representing a Finder folder should use the same icon as visible in the Finder, i.e. on Leopard a 
10.5 folder icon, and on Tiger a 10.4 folder icon. This is easy to achieve by asking NSWorkspace for the 
folder icon. It has the additional benefit that icons for special folders like Music, Pictures, and Movies 
folders are respected.

The attached patch takes care of this.

Cheers,
Peter