This is a tiny change with a big explanation.
I took out the +initialize that called [NSWorkspace sharedWorkspace] to populate a couple of static variables, and put the code to populate the statics inline where they were used.

Why?  It turns out that on Tiger, +initialize methods are getting called from "start" before main gets called.  And main is the method that is responsible for  creating the appropriate subclass of NSApplication from the info.plist.  So with the way it was, [NSWorkspace sharedWorkspace] was instantiating an instance of NSApplication -- not the appropriate subclass.

This will help for any apps that subclass NSApplication and that need to run on Tiger.