Tiger Compatibility -- iMedia will not work under Tiger without this! :-)
In iMBMusicFolder and iMBMoviesFolder, we were making use of [QTMovie movieTypesWithOptions:QTIncludeAllTypes] which is a QuickTime 7.2 API.  Tiger seems to have an older version of QuickTime.  (This would explain why building on Leopard doesn't show problems.)

Since all the logic below deals with UTIs, I decided the best way to do this was convert the file types (extensions or 'XXXX' codes) into UTIs and then continue with the comparison.  

So this won't really change anything for Leopard; but it will bring back decent functionality back to Tiger.



Also, for minimal systems with no iPhoto, use Preview's icon for the toolbar.