A minor checkin to deal with some compiler warnings from the Clang static analyzer -- leaks, bad arguments passed to NSLog, etc.
Note the use of NSMakeCollectable, to prevent leaks for when this code is compiled under the GC flavor of Objective C.  There is probably some real effort to ensure that this is GC-friendly, but this is a first start, by making some core foundation non-autoreleased objects be autoreleaed in the toll-free-bridged world.  However, we define it as a no-op for backwards compatibility with pre-Leopard SDKs.