A tiny but important fix for UKKQueue that prevents it from watching the same path multiple times.
* * * * MAJOR KUDOS TO RAINER BROCKERHOFF FOR FINDING THIS PUPPY!

This should fix that instability problems that some were reporting when live library updating was enabled.  It turns out that the same path was getting watched many many times, and you would run out of file descriptors.  Now with even a pretty big iPhoto library you are not likely to hit your open file limitation.

BTW if you want to "up" your open file allocation, here is some code to set that (to a particular number or the max) ... put this first in your main() function.

	struct rlimit limit = {0,0};
	getrlimit(RLIMIT_NOFILE,&limit);
	if (limit.rlim_cur<limit.rlim_max) {
		limit.rlim_cur = 10000; // limit.rlim_max;
		setrlimit(RLIMIT_NOFILE,&limit);
	}