Dutch Localizable.strings contributed by NeoOffice.org
(We will need to get more resources localized to be truly useful)


From: Patrick Luby <pluby@________>
Date: Tue, Sep 23, 2008 at 12:56 PM
Subject: License of Dutch localizations for iMedia
To: dan94501@gmail.com
Cc: openstep@________, Oscar van Vliet <________>


Hello,

I am sending you this e-mail because you are listed as the iMedia
project owner in Google Code.

I and Ed Peterlin (cc'd as openstep@________) are the founders of
NeoOffice.org and we use the iMedia browser within our open source product.

Many of our users really like your work and, so far, ________
has translated the Localizable.strings file to Dutch. Since our license
is GPL and is inconsistent with your license, we want to formally
license the translations back to your project under the iMedia source
code so that you can include them in your project.

If this sounds agreeable to you, then feel free to commit the attached
Localizable.strings and use this e-mail as proof of the our licensing to
you.

Thank you,

Patrick

--