Sharking dismal performance experienced while iMedia parses an Aperture library, the majority of time is spent in
- (NSArray *)recursiveAttributesForKey:(NSString *)key filterKey:(NSString *)filter excludingSet:(NSMutableSet *)alreadyAdded

just dealing with the enumerations. As conceptually wonderful as enumerations are, using them in a recursive method quickly builds up a crazy amount of overhead. I've surgically replaced the enumerators in this method with a standard for loop and the problem has abated.

I tried to change as little of the original code as possible: just enough so that we can comfortably ship iMedia Browser. We should have a deeper look at this class and see what other improvements can be made.