Patch from daniel jalkut:
I'm using the "selection changed" notifications to keep account of
which items in the media browser I will act on when the user clicks a
separate button in my interface, or when they double click on the
selected items.

This works great in general, but I've noticed a behavior of the
double-
click handling in iMBPhotosView.m, that causes problems for me.

In particular, when an item is double-clicked, iMedia intentionally
sends out a notification falsely indicating that there is no
selection.