Preliminary combo view with asynchronous loading.
This is largely based on AnimatedTableView sample source from Apple.  It's mostly functional except the notification of the loading finished, which I need to rewrite in Leopard APIs instead of the Snow Leopard APIs that they gave us.

What I'm hoping to do is to properly have the asynchrnous loading stuff go into the dynamic table view, and the multiple sub-cell stuff go into the combo table view.  Then in theory we could have the list view still load up tiny thumbnails, for instance.  This isn't quite right though.

I've written an IKImageBrowserview-like client for the image data source so that we can get synchronous or asynchronous thumbnails.  That has been started but there is probably a lot of optimiztion to do, and I need to make sure it works with ImageCapture.

Still, it's a start.