Implement audio-stopping functionality in the audio view controller. I decided to implement this using the play button as both the play and stop button, which is exactly analogous to how the Apple iLife browser handles it. You'll see whenever playback is happening the triangle in the play button turns blue (highlighted), and from this state, pushing the button again will stop playback.
I also took the liberty of tweaking the display of the playback button's triangle
so that it is more centered in the button. It's kind of hack that this was achieved
by changing the image position in IB to "under the title", but it seems to work, and
should be relatively stable.

--HG--
extra : convert_revision : svn%3A0a3f77c4-9235-0410-95f4-d138c4647c51/trunk%40871