Get Bookmark items working so that double-clicking them causes them to be opened as expected in the default browser/schemehandler.
To achieve this, I altered the approach of IMObjectPromise, as discussed with Peter B, so that it vends an array of urls instead of an array of file paths. This allows us to use the same type object promise for web urls as we use for local files, regarding them as "local" in the sense that they don't require any extra processing or asychnronous loading in order to fulfill a user demand with them.

I tested with Flickr, Aperture, iPhoto, and iTunes sources to confirm that dragging from the iMedia browser to the desktop still works as expected,
and that double-clicking items opens them up in the appropriate application, as before.

Double-clicking bookmarks now typically causes them to open the URL in e.g. Safari.