Revise IMBObjectViewController's relationship with the pasteboard, as described on the mailing list.
The idea here in a nutshell is to take advantage of 10.6's new multiple item pasteboard, so that things like multiple URLs can be dragged to a destination at once.

I implemented the functonality in such a way that clients using the 10.5-style pasteboard APIs should notice no change in behavior. They can still request an object promise or an array of filenames.

To achieve the desired functionality, I modified IMBObjectPromise so that it keeps track of objects and whether they have been fulfilled or not yet. When they are fulfilled, the local URL is stored in a dictionary keyed by the object itself. This lets us use the complete object promise for all objects as the object promise for any given item on the pasteboard. The granular fulfillment information might eventually be useful for allowing only the items that are desired to be downloaded.

Other changes of note in this checkin:

- IMBURLDownloadOperation now accepts offers to store a delegateReference attribute, which a client can use for whatever purpose.

- URL dragging now works! (It didn't work at all previously, and now works with support for multiple items on 10.6).

- The custom pasteboard type for object promises was modified so that it uses a format that is legal for UTI (important for the 10.6 style pasteboard stuff).

- IMBObjectViewController subclasses must now take care to advertise whether the types of cargo they place on the pasteboard are local files or not. This is required so that the correct types can be declared in the generic pasteboard handling code in IMBObjectViewController.

I tested these changes on 10.6 and 10.5, and confirmed that e.g. for 10.5-style pastboard API users, the contents of the pasteboard can still be retrieved as before.