Some of my IMBObjects have metadata that is not strictly NSString, but I still want filtering to work on these objects and their metadata. I hope I'm not being too presumptuous by assuming it's OK to "abuse" the metadata system in this manner, but it seems to work fine. Anyway, I thought that it might be an issue that comes up in the future for others, so I enhanced the array controller to support custom metadata values in the search filtering procedure.
The old contract was: all metadata values must be NSString (implicitly, or they would throw exceptions when messaged with "lowercaseString").

The new contract is: all metadata values must be NSString or else implement a custom search filtering method "- (BOOL) matchesSearchFilterString:"

Let me know if there are any concerns about the choice of method name, or about this practice of expanding the possible values of metadata. I'm happy to adjust as concerns warrant.