Setting SDKRoot back to 10.5, as Daniel Jalkut had done in r953.  Peter B had put it back to 10.6 when moving things to separate .xcconfig files, but I don't think it was intentional.
I think that in order to be 10.5 compatible, we have to have SDKRoot set to 10.5 so that we link to libcrypto.0.9.7.dylib (which is in Leopard) not libcrypto.0.9.8.dylib (which is in Snow Leopard).  Unfortunately I don't see any way to link to just 'libcrypto.dylib' and get the latest version.

If anybody thinks that this isn't a good solution (e.g. it would break other things) let's figure out a way to keep things 10.5 compatible! :-)

--HG--
extra : convert_revision : svn%3A0a3f77c4-9235-0410-95f4-d138c4647c51/trunk%401024