Change the DYLD compatibility and current versions to 2, to match the current version of the ObjectiveFlickr framework.
Having the versions set to 1 caused iMedia to link against the framework with a minimum requirement of version 1, but there are binary differences between the older and newer ObjectiveFlickr, so iMedia will fail to resolve all expected symbols if it encounters a version 1 of the framework.

I ran into this problem with MarsEdit where a customer had an older copy of ObjectiveFlickr installed in /Library/Frameworks. Since I don't ship ObjectiveFlickr with MarsEdit, it's normally fine if the library is not found, because it's a weak import. But if DYLD finds the user's 1.0 version of the framework, it tries to resolve all of iMedia's weak symbols against it, and fails to find expected symbols such as e.g. OFFlickrThumbnailSize.