Patch from Gideon @ Novamind (submitted a while ago):
IMBComboTextCell: Fixed the calculation of the title rect.
IMBLinkViewController: Fixed a problem where it tried to display an icon view for hyperlinks when the user switched to that panel
IMBNodeViewController: Made the registration as a drop destination able to be controlled by a delegate for situations where you may not want to accept additional folders etc.
IMBOutlineView: Made the display of the prompt for dropping folders dependent on registration for drags.
IMBPanel: New class which overrides the normal close method to order out the window. This is to avoid a bug in the NSImageBrowserView where if there are selections when the window is closed, there are error messages printed to the console log. Not too much of an issue in the case where it is printing only for one selected object, but I have encountered a situation where it printed an error message for every image in my library, which stalled my program for well over a minute.
IMBPanel.xib: Changed the class of the panel to IMBPanel (There appears to be an issue where some outlets are missing, but I didn't have a chance to look into that)
NSString+iMedia: Made it create a new date formatter each time instead of using a static one seeing as I was getting exceptions thrown when I reused the static one (appears to be some bug in Apple's code)

--HG--
extra : convert_revision : svn%3A0a3f77c4-9235-0410-95f4-d138c4647c51/trunk%401093