Minor update to ObjectiveFlickr which is in sync with the version that has been stable for several months now on GitHub.
I'm thinking that perhaps it might be a good idea to NOT include ObjectiveFlickr in the imedia package, and require it to be checked out from GitHub.  However that might play havoc with the build process, so perhaps this step should not be taken lightly.

--HG--
extra : convert_revision : svn%3A0a3f77c4-9235-0410-95f4-d138c4647c51/trunk%401118