Trying to debug drag of a single item from flickr into the tester app's window.  Sometimes, not very often, it messes up, and there are two progress windows that open up.
It seems that -[IMBObjectViewController pasteboard:provideDataForType:] is getting called twice, back-to-back.  One of the drags fails, logging this to console:  "Looked for URLs on the pasteboard, but found none"  By logging a backtrace on pasteboard:provideDataForType: (and auto-continuing, to not interrupt the flow) I can see two slightly different backtraces.  How is the method being called twice when it's supposed to wait for the result? It's like the run loop it calls when it waits for a drag is letting another drag get posted!

Also a mystery: why is the cancel button not working on the progress dialog?  Well, I got it to work one time only.  Never since then.

Can somebody (Peter, somebody else?) try debugging this for a while and see if you have any luck?