IMBObjectPromise now checks to see if downloadable file already exists.
Right now all we do is log that .. but next step after lunch will be to somehow avoid doing the whole download process, and skip to the part where we present the final file.  (This might be tricky!)

I WONDER: should we have a way to force a new download if the file exists?  e.g. option-drag or something to disallow using the cached version?

Cleaned up a few compiler warnings too.