Async downloads now have 3 modes for counting progress.
1 file: don't do a pre-fetch of HEAD; just start downloading and as soon as we have the expected content length, we set up the progress meter

a small number of files: Asynchronously request the HEAD so we get an accurate expected byte count before we begin downloading, so that progress bar increments properly as chunks of data come in

larger number of files (currently set at 8 or more): Just increment progress bar as each file comes through. It's not as smooth of a progress update (especially with parallel downloads coming in) but it avoids delays in pre-counting a lot of files.

--HG--
extra : convert_revision : svn%3A0a3f77c4-9235-0410-95f4-d138c4647c51/trunk%401211