Change "song"/"songs" to "track"/"tracks" as in iMedia 1.  This means that we may want to change the translations.
Updates to English.lproj, translations will be affected:

DELETED:
Date
Size
Type

ADDED:
Comment
Owner
Tags



CHANGED:
IMBAudioViewController.countFormatPlural   (now: tracks)
IMBAudioViewController.countFormatSingular  (now: track)

IMBObject.untitled --- was "Untitled" but changed to "untitled".  This is because it is not for an object called "Untitled" (with a capital letter) but just that it is untitled, so I think that in a list, it is better to have no capital letter.  This might not be the same in other languages :-)