debug logging some KVO ...
In the movies view of itunes movies.  I had switched to another app, then switched back.  Thumbnails weren't showing.   Scrolled, and got this over and over.

Cannot remove an observer <IMBMovieViewController 0x102058710> for the key path "imageRepresentation" from <IMBMovieObject 0x12d654a70> because it is not registered as an observer.

So I'm logging each time we add and remove observation, in case thsi comes up again.