mostly some cleanup of a method name that a global search and replace had messed up.
Also, in folder parser, check to see if a sub-folder has any sub-folders of its own, so it shows a disclosure triangle more appropriately.  The scan bails out after 100 files -- in case you have a huge directory -- with an assumption that there may be subfolders there.  This may lead to some false positives but it will keep the system zippy.